Refactor project structure and modularize services
- Moved user, product, authentication, and index logic into services folder
